requirements.txt
langchain_core
python-dotenv
langchain_google_genai
streamlit
flask
python-docx